requirements.txt
typer==0.12.5
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
anthropic==0.39.0
rich==13.9.2
jinja2==3.1.4
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
gspread==6.1.2
google-auth==2.35.0
